added the configured project
