+++ Code/SSI-Wallet/ssi_mobile_wallet__backend/.env
@@ -1,0 +1,1 @@
+AUTHTOKEN="CiVodHRwczovL3RyaW5zaWMuaWQvc2VjdXJpdHkvdjEvb2Jlcm9uElYKK3Vybjp0cmluc2ljOndhbGxldHM6ejhvYVZnSjJ1VERGQ1NnaVFXeTJSQ3QiJ3Vybjp0cmluc2ljOmVjb3N5c3RlbXM6dm9tc21vYmlsZXdhbGxldBowl0wYRsPZxU_pgvV656hqcPUxxx-omOSHMZigJGVWPa3I8Ow1Sk7SxGGsiKkP58GUIgA"
