--- Code/SSI-Wallet/ssi_mobile_wallet__backend/nodemon.json
@@ -1,6 +1,0 @@
-{
-    "watch": ["src"],
-    "ext": ".ts,.js",
-    "ignore": [],
-    "exec": "npx ts-node ./src/index.ts"
-  }
