+++ Code/SSI-Wallet/ssi_mobile_wallet__backend/package.json
@@ -4,23 +4,22 @@
   "description": "",
   "main": "index.js",
   "scripts": {
-    "start:dev": "npx nodemon",
+    "start": "nodemon --exec ts-node server/index.ts",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "keywords": [],
   "author": "",
   "license": "ISC",
-  "devDependencies": {
-    "@trinsic/trinsic": "^1.10.0",
+  "dependencies": {
+    "@trinsic/trinsic": "1.10.0",
+    "dotenv": "^16.0.3",
     "@types/express": "^4.17.17",
-    "@types/node": "^18.15.11",
     "express": "^4.18.2",
+    "uuid": "^9.0.0"
+  },
+  "devDependencies": {
     "nodemon": "^2.0.22",
     "ts-node": "^10.9.1",
-    "typescript": "^5.0.3"
-  },
-  "dependencies": {
-    "@trinsic/trinsic": "^1.10.0",
-    "dotenv": "^16.0.3"
+    "typescript": "^5.1.3"
   }
 }
